+++ Metodos_de_classes_e_funcoes_criadas.cpp
@@ -1,13 +1,52 @@
+#include "Declaracao_de_classes_e_funcoes_criadas.h"
 #include <cstring>
 #include <utility>
 #include <vector>
 
-#include "FuncoesAuxiliares.h"
-#include "Metodos_arqbin.h"
-#include "Classes_metodos_grafo.h"
-
-#include<iostream>
+#include "FuncoesPadrao.h"
 using namespace std;
+
+Cabecalho::Cabecalho(FILE* arquivo) {
+    fread(&status, 1, 1, arquivo);                 // Lê 1 byte para 'status'
+    fread(&topo, sizeof(int), 1, arquivo);         // Lê 4 bytes para 'topo'
+    fread(&proxRRN, sizeof(int), 1, arquivo);      // Lê 4 bytes para 'proxRRN'
+    fread(&nroRegRem, sizeof(int), 1, arquivo);    // Lê 4 bytes para 'nroRegRem'
+    fread(&nroPagDisco, sizeof(int), 1, arquivo);  // Lê 4 bytes para 'nroPagDisco'
+    fread(&qttCompacta, sizeof(int), 1, arquivo);  // Lê 4 bytes para 'qttCompacta'
+}
+
+Dado_bin::Dado_bin(FILE *arquivo) {
+    fread(&removido, 1, 1, arquivo);                       // Lê 1 byte para 'removido'
+    fread(&encadeamento, 4, 1, arquivo);                   // Lê 4 bytes para 'encadeamento'
+    fread(&populacao, 4, 1, arquivo);                      // Lê 4 bytes para 'populacao'
+    fread(&tamanho, 4, 1, arquivo);                        // Lê 4 bytes para 'tamanho'
+    fread(&unidadeMedida, 1, 1, arquivo);                  // Lê 1 byte para 'unidadeMedida'
+    fread(&velocidade, 4, 1, arquivo);                     // Lê 4 bytes para 'velocidade'
+    fread(&variavel, 1, 142, arquivo);                     // Lê 142 bytes para 'variavel'
+
+    //cout << variavel << endl;
+}
+
+
+
+Especie::Especie(FILE* arquivo) {
+    Dado_bin auxiliar(arquivo);
+    char *lin = strdup(auxiliar.variavel);
+
+    populacao = auxiliar.populacao;
+
+    nome = strsep(&lin, "#");
+    especie = strsep(&lin, "#");
+    habitat = strsep(&lin, "#");
+    dieta = strsep(&lin, "#");
+    tipo = strsep(&lin, "#");
+    presa = strsep(&lin, "#");
+
+    grau_entrada = 0;
+    grau_saida = 1;
+    grau = 1;
+}
+
 Presa::Presa(int populacao, string nome) {
     populacao_do_predador = populacao;;
     nome_da_presa = std::move(nome);
@@ -113,6 +152,7 @@ Grafo::Grafo(FILE* arquivo) {
 }
 
 
+
 void Cria_grafo_e_exibe() {
 
     Grafo g = Cria_grafo();
@@ -129,8 +169,6 @@ Grafo Cria_grafo() {
     return Grafo (arq_bin);
 }
 
-//METODO USADO NA FUNCAO 11///////////////////////////////////////////////////////////////////////
-
 void Grafo::exibe_grafo() const {
 
     if(numero_de_vertices==-1)
@@ -146,9 +184,93 @@ void Grafo::exibe_grafo() const {
     }
 }
 
+void Exibe_predadores() {
+    Grafo g = Cria_grafo();
+
+    if(g.numero_de_vertices==-1) //se houver problema na criacao do grafo
+    {
+        cout << ERRO_PADRAO;
+        return;
+    }
+
+    int numero_de_buscas = 3;
+    cin >> numero_de_buscas;
+
+    fflush(stdin);
+
+    string input;
+    string busca_atual;
+    vector<string> busca;
+    bool dentro_aspas = false;
+
+    getline(cin, input);
+
+    for (const char c : input) {
+        if (c == '"') {
+            if (dentro_aspas) {
+                busca.push_back(busca_atual);
+                busca_atual.clear();
+            }
+            dentro_aspas = !dentro_aspas;
+        }
+        else if (dentro_aspas) {
+            busca_atual += c;
+        }
+    }
+
+    for(int i=0; i<numero_de_buscas; i++) {
+        string saida;
+
+        for(const auto& x:g.vertices) {
+            for(const auto& y:x.presas) {
+                if (y.nome_da_presa==busca[i]) {
+                    saida += x.predador.nome + ", ";
+                }
+            }
+        }
+        if(!saida.empty()) {
+            saida.erase(saida.size()-2,2);
+            cout << busca[i] << ": " << saida<<endl << endl;
+        }
+        else {
+            cout << ERRO_REGISTRO << endl;
+        }
+    }
+}
 
 
-//METODOS USADOS NA FUNCIONALIDADE 13///////////////////////////////////////////////////////////
+//FUNCIONALIDADE 13 - DFS
+
+//a funcao Analisa_Conexoes() eh a funcao de pesquisa em profundidade do grafo
+//essa funcao cria o grafo e chama a rotina Grafo::Profundidade para retornar a qtd
+//de componentes conexos dentro do grafo para entao retornar o resultado
+
+//as duas principais funcoes utilizadas aqui são Profundidade e Profundidade_recursao, explicadas a seguir
+
+void Analisa_conexoes() {
+    Grafo g = Cria_grafo();      //cria o grafo
+
+    if(g.numero_de_vertices==-1) //se houver problema na criacao do grafo, da erro e nao executa a funcao
+    {
+        cout << ERRO_PADRAO;
+        return;
+    }
+
+    int componente;     
+    componente=g.Profundidade();    //chama a funcao de pesquisa em profundidade, necessaria para saber se um grafo eh ou nao eh fortemente conexo
+                                    //essa funcao retorna a qtd de componentes conexos do grafo
+
+    //se o numero de componentes conexos for igual ao nro de vertices o grafo eh fortemente conexo
+    //assim partindo de qualquer vertice eh possivel chegar em todos os outros vertices do grafo
+
+    //caso isso nao seja verdade ele nao eh fortemente conexo, mas ainda sim podem ter vertices que tenham a capacidade
+    //de chegar em todos os outros vertices do grafo, so nao sao todos
+    if(componente==g.numero_de_vertices)
+        cout << "Sim, o grafo e fortemente conexo e possui 1 componente";
+    else
+        cout << "Nao, o grafo nao e fortemente conexo e possui " << componente << " componentes";
+    
+}
 
 // funcao principal da pesquisa em profundidade
 //a funcao realiza a pesquisa em profundidade de forma recursiva em todos os vertices do grafo
@@ -230,7 +352,48 @@ void Grafo::Profundidade_recursao(Predador vertice, int x, vector<vis> &visitado
     }
 }
 
-//METODO USADO NA FUNCIONALIDADE 14////////////////////////////////////////////////////////////////
+//FUNCIONALIDADE 14 - DIJKSTRA
+
+//a funcao realiza um loop de pesquisas que recebe
+//um vertice de inicio e um alimento alvo e determina o peso
+// minimo do caminho que parte do inicio e vai ate o alvo
+//isso eh feito com o algoritmo de dijkstra, que segue abaixo
+
+void Relacao_presa_predador(){
+
+    Grafo g = Cria_grafo();      //cria o grafo
+
+    if(g.numero_de_vertices==-1) //se houver problema na criacao do grafo, da erro e nao executa a funcao
+    {
+        cout << ERRO_PADRAO;
+        return;
+    }
+
+    int pesquisas;          //nro de pesquisas que serao feitas
+    int peso_caminho;       //variavel de retorno da pesquisa, indica o peso do caminho
+    char n_presa[91];       //presa da pesquisa
+    char n_predador[91];    //predador da pesquisa
+
+    cin >> pesquisas;   //quantas pesquisas serao feitas
+
+    for(int i=0;i<pesquisas;i++)    //loop das pesquisas
+    {   
+        scan_quote_string(n_predador);  //le o primeiro e ultimo vertice da pesquisa
+        scan_quote_string(n_presa);
+
+
+        peso_caminho=g.dijkstra(n_predador,n_presa);//executa a rotina do algoritmo de dijkstra no grafo
+                                                    //funcao acha o menor caminho entre dois vertices
+                                                    //funcao retorna -1 se nao houver caminho entre os dois
+
+        //ao fim da pesquisa atual checa se teve ou nao caminho e exibe o resultado
+        if(peso_caminho==-1)    
+            cout << "CAMINHO INEXISTENTE";
+        else 
+            cout << peso_caminho;
+        cout << "\n";
+    }
+}
 
 //o algoritmo recebe o vertice inicial e o alvo, o algoritmo guloso
 //funciona da maneira determinada nos slides da AULA17
@@ -239,35 +402,29 @@ void Grafo::Profundidade_recursao(Predador vertice, int x, vector<vis> &visitado
 //no vetor D vale -1
 
 int Grafo::dijkstra(char n_predador[91],char n_presa[91])
-{   
-    int i;
+{
     int peso=0;
     vector<dij> D;  //vetor D que armazena os pesos para chegar ate cada vertice partindo de um vertice de predador
     dij temp;             // vetor auxiliar
     //inicializa o vetor auxiliar D
 
-    // inicializa o vetor auxiliar que sera utilizado na pesquisa, basicamente os pesos para chegar ao vertice iniciam como -1 exceto o inicial
-    // e tambem eh destacado o nome da especie
+    // inicializa o vetor auxiliar que sera utilizado na pesquisa, basicamente os pesos para chegar ao vertice iniciam como -1
+    // e tambem eh destacado o nome da especie, se nao houver caminho que chegue a algum vertice
+    temp.peso = -1;
     for (const auto &x : vertices)
     {
-        temp.peso = -1;
         temp.nome = x.predador.nome;
-        if (x.predador.nome == n_predador)  //peso do vertice de partida eh zero
-            temp.peso=0;
         D.push_back(temp);
     }
-
-
-
-
+    
+        
 
 
 
     //retorna o peso do caminho para chegar na presa/alimento a partir do predador dado
-    for (i = 0; i < numero_de_vertices; i++) 
+    for (int i = 0; i < numero_de_vertices; i++) 
         if (D[i].nome == n_presa)
-            break;
-    return D[i].peso;
+            return D[i].peso;
 }
